Add failing feature for fixture directory override
For projects configured with rspec as the test framework, the railtie
overrides any user-specified fixture directory that might be included in
config/application.rb.

This patch modifies the required behaviour so that if such configuration
is detected, it is not overridden. It also expects the auto-added
configuration to match the generator's default of 'test/factories'.