Allow Rails 3.1 to select the 'turn' gem it wants to use
Rails 3.1.x has code which checks which version of ruby is being run,
and based on the result, selects a version of turn.

For 1.9.3+, it selects "~> 0.8.3"; for everything else, it selects "0.8.2".

If we remove turn from the Appraisals, but have travis install both versions
that rails might potentially need, the in-generated-app Gemfile building should
work correctly.