Skip Spring version 2.1.1 (#389)
* Skip Spring version 2.1.1

Changes introduced in spring 2.1.1
(https://github.com/rails/spring/pull/621) are breaking some tests. That
change was reverted in https://github.com/rails/spring/pull/629, but
hasn't been released yet. Until #629 is released, this PR skips Spring
version 2.1.1.

* Bump to latest standard to match CI

We are using the latest standard on CI. This bumps the version in the
dev Gemfile to match, and fixes one violation.

Co-authored-by: Daniel J. Colson <daniel.colson@hey.com>