Update all dev dependencies
Ran `bundle update` so the dev Gemfile is using the latest versions of
everything. As long as the tests pass, this should be fine.

This commit also adds the missing Appraisal from
https://github.com/thoughtbot/factory_bot_rails/pull/410 (the PR
description mentioned that it was missing, but it didn't register for me
when I was reviewing the PR).