Be clear about supported Rails and Ruby versions
[Ruby 2.7 and older] is end of life.

[Rails 6.0 and older] is end of life.

Rails 6 [requires] Ruby >= 2.5, Rails 7 requires Ruby >= 2.7.

[Ruby 2.7 and older]: https://www.ruby-lang.org/en/downloads/branches/
[Rails 6.0 and older]: https://guides.rubyonrails.org/maintenance_policy.html
[requires]: https://guides.rubyonrails.org/upgrading_ruby_on_rails.html#ruby-versions

---

Don't require debug. This isn't needed for tests. But moreover, it
causes an error on Ruby 3.0.

It opens with:

``ruby
if $SAFE > 0
  STDERR.print "-r debug.rb is not available in safe mode\n"
  exit 1
end
```

But `$SAFE` is not set, so we get `nil > 0`, and an error.

```
% ruby -rdebug -e 'puts "OK"'
/home/mike/.rubies/ruby-3.0.6/lib/ruby/3.0.0/debug.rb:9:in `<top (required)>': undefined method `>' for nil:NilClass (NoMethodError)
        from <internal:/home/mike/.rubies/ruby-3.0.6/lib/ruby/3.0.0/rubygems/core_ext/kernel_require.rb>:85:in `require'
        from <internal:/home/mike/.rubies/ruby-3.0.6/lib/ruby/3.0.0/rubygems/core_ext/kernel_require.rb>:85:in `require'
```

This is not an issue on Ruby 3.1.

---

Ruby 3.1 + Rails 7.1 leads to an issue around the `error_highlight`
library. Rather than debug that, we can skip that combination and hope
for the best.