Fix development dependencies (#526)
Some changes in our dependency tree have led to broken builds. For
example, `concurrent-ruby`, which Rails uses, stopped requiring `logger`
on version 1.3.5, which was fixed in Rails 7.1. Another change is that
Ruby decreased its standard library, removing some preloaded gems, such
as `mutex_m`.

This commit updates the dependencies to the latest supported version,
limits the `concurrent-ruby` version to `< 1.3.4` in Rails versions that
don't require `logger`, and adds `mutex_m` to the list of development
dependencies.