Limit `push` builds to the `main` branch
Commits to a Pull Request branch have been triggering a build twice.
That is happening because a `push` to a pull request is both a `push`
and a `pull_request` activity.

This commit fixes this issue by limiting the `push` trigger to the
`main` branch, ensuring that merged Pull Requests will run linting and
tests on the main branch, and that any Pull Request activity will also
trigger a build.