Add latest Ruby and Rails versions to the build (#529)
* Add latest Ruby and Rails versions to the build

The Rails versions had been added to the Appraisals file in previous
PRs, but GitHub Actions didn't run with them because the `build.yml`
file hadn't been changed.

* Add `drb` gem to Rails 6.1 and 7.0

The `drb` gem is not a default gem on Ruby 3.4 anymore, so it needs to
be on the Gemfile. Rails 7.1 and above already include it as a
dependency.